requirements.txt
tda-api
selenium
